authed_mcp_demo/requirements.txt
authed
authed-mcp
fastapi
uvicorn
httpx
asyncio
pydantic
python-dotenv
# The official 1Password SDK for Python
onepassword-sdk
